requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.0

# DB & ORM
sqlalchemy==2.0.27
psycopg2==2.9.6
PyMySQL==1.1.1
pyodbc==5.2.0

# Redis
redis==5.2.1

# Pydantic, data
pydantic==2.10.6
pandas==2.2.0
numpy==1.26.4

# Plotting
plotly==5.24.1
matplotlib==3.10.0

# NLP
transformers==4.48.1
torch==2.5.1
sentence-transformers==3.4.0
scikit-learn==1.6.1
sentencepiece==0.2.0

# OpenAI (if you want ChatGPT / GPT-4)
openai==1.60.2


# Additional
setuptools==75.8.0
wheel==0.45.1
